Move uploadBlockBlob to its own file
